Move report calculators to calculators.py
There's no reason why these should be in models.py.
