+++ dmt/report/calculators.py
